refactor: split large test file into focused test files
- Fixed fragile HTTP method selection in tests (addressing PR review)
- Split 3367-line group_test.go into 6 focused test files:
  * group_test.go (560 lines) - core middleware and group tests
  * mount_test.go (584 lines) - mount operations and nested groups
  * routing_test.go (925 lines) - route handling and HTTP methods
  * notfound_test.go (202 lines) - custom 404 handler tests
  * fileserver_test.go (659 lines) - static file serving tests
  * pathparams_test.go (519 lines) - path parameters and URL encoding
- All 43 test functions preserved and passing
- Better organization for maintainability and parallel execution

Related: #22
